+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `optimize` option to `build.zig`.
+
 ## [0.1.0] - 2025-08-22
 
 Initial release.
@@ -13,4 +19,5 @@ Initial release.
   `toml.parseWithDiagnostics`.
 - All TOML value types.
 
+[Unreleased]: https://github.com/anttikivi/toml-zig/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/anttikivi/toml-zig/releases/tag/v0.1.0